% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\name{create_keypair}
\alias{create_keypair}
\alias{read_private_key}
\alias{read_public_key}
\title{Create keypair}
\usage{
create_keypair(path, password, quiet = FALSE, name = "id_encryptr")

read_private_key(path)

read_public_key(path)
}
\arguments{
\item{path}{Path to the directory to store the key.}

\item{password}{What do we do about passwords?  Options are
\code{FALSE} for no password, \code{TRUE} for prompting for a
password, or a string value for a password (which will end up in
things like your history so be careful).}

\item{quiet}{Suppress printing of informative messages.}

\item{name}{Leave this be for now.}
}
\description{
Create a pair of public and private keys, possibly with password
protection.  If you're trying to do the data workflow thing, use
\code{\link{data_user_init}} instead, which wraps this but puts
the key in the place the rest of the functions there will look.
}
\details{
The functions \code{read_public_key} and \code{read_private_key}
will read the public and private key stored at \code{path}.  If
the private key is password-protected then you will be prompted
for a password.
}

