% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyphr.R
\docType{package}
\name{cyphr-package}
\alias{cyphr}
\alias{cyphr-package}
\title{High Level Encryption Wrappers}
\description{
Encryption wrappers, using low-level support from sodium and openssl.
}
\details{
It is \emph{strongly} recommended that you read \emph{both} vignettes before
attempting to use \code{cyphr}.
\itemize{
\item \href{https://docs.ropensci.org/cyphr/articles/cyphr.html}{introduction};
in R: \code{vignette("cyphr", package = "cyphr")}
\item \href{https://docs.ropensci.org/cyphr/articles/data.html}{data vignette};
in R: \code{vignette("data", package = "cyphr")}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/cyphr}
  \item \url{https://docs.ropensci.org/cyphr/}
  \item Report bugs at \url{https://github.com/ropensci/cyphr/issues}
}

}
\author{
Rich FitzJohn (rich.fitzjohn@gmail.com)
}
