% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyphr.R
\docType{package}
\name{cyphr}
\alias{cyphr}
\alias{cyphr-package}
\title{High Level Encryption Wrappers}
\description{
Encryption wrappers, using low-level support from sodium and openssl.
}
\details{
It is \emph{strongly} recommended that you read \emph{both} vignettes before attempting
to use \code{cyphr}.
\itemize{
\item \href{https://ropensci.github.io/cyphr/articles/cyphr.html}{introduction} (in R: \code{vignette("cyphr", package = "cyphr")})'
\item \href{https://ropensci.github.io/cyphr/articles/data.html}{data vignette} (in R: \code{vignette("data", package = "cyphr")})'
}
}
\author{
Rich FitzJohn (rich.fitzjohn@gmail.com)
}
