% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{data_key_read}
\alias{data_key_read}
\title{Read a public key}
\usage{
data_key_read(path = NULL)
}
\arguments{
\item{path}{Either \code{NULL} or a directory name to read your
public key, or a full filename of a file to read a different
key.}
}
\description{
Read a public key
}

