% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt_data.R
\name{encrypt_data}
\alias{decrypt_data}
\alias{decrypt_file}
\alias{decrypt_object}
\alias{decrypt_string}
\alias{encrypt_data}
\alias{encrypt_file}
\alias{encrypt_object}
\alias{encrypt_string}
\title{Encrypt and decrypt data and other things}
\usage{
encrypt_data(data, dest, config)

encrypt_object(object, dest, config)

encrypt_string(string, dest, config)

encrypt_file(path, dest, config)

decrypt_data(data, dest, config)

decrypt_object(data, config)

decrypt_string(data, config)

decrypt_file(path, dest, config)
}
\arguments{
\item{data}{(for \code{encrypt_data}, \code{decrypt_data},
\code{decrypt_object}, \code{decrypt_string}) a raw vector with
the data to be encrypted or decrypted.  For the decryption
functions this must be data derived by encrypting something or
you will get an error.}

\item{dest}{The destination filename for the encrypted or
decrypted data, or \code{NULL} to return a raw vector.  This is
not used by \code{decrypt_object} or \code{decrypt_string} which
always return an object or string.}

\item{config}{A \code{encryptr_config} object describing the
encryption approach to use.}

\item{object}{(for \code{encrypt_object}) an arbitrary R object to
encrypt.  It will be serialised to raw first (see
\code{\link{serialize}}).}

\item{string}{(for \code{encrypt_string}) a scalar character
vector to encrypt.  It will be converted to raw first with
\code{\link{charToRaw}}.}

\item{path}{(for \code{encrypt_file}) the name of a file to
encrypt.  It will first be read into R as binary (see
\code{\link{readBin}}).}
}
\description{
Encrypt and decrypt raw data, objects, strings and files.  The
core functions here are \code{encrypt_data} and
\code{decrypt_data} which take raw data and decrypt it, writing
either to file or returning a raw vector.  The other functions
encrypt and decrypt arbitrary R objects (\code{encrypt_object},
\code{decrypt_object}), strings (\code{encrypt_string},
\code{decrypt_string}) and files (\code{encrypt_file},
\code{decrypt_file}).
}
\examples{
key <- sodium::keygen()
cfg <- config_sodium_symmetric(key)
# Some super secret data we want to encrypt:
x <- runif(10)
# Convert the data into a raw vector:
data <- serialize(x, NULL)
data
# Encrypt the data; without the key above we will never be able to
# decrypt this.
data_enc <- encrypt_data(data, NULL, cfg)
data_enc
# Our random numbers:
unserialize(decrypt_data(data_enc, NULL, cfg))
# Same as the never-encrypted version:
x

# This can be achieved more easily using `encrypt_object`:
data_enc <- encrypt_object(x, NULL, cfg)
identical(decrypt_object(data_enc, cfg), x)

# Encrypt strings easily:
str_enc <- encrypt_string("secret message", NULL, cfg)
str_enc
decrypt_string(str_enc, cfg)
}

