% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{encrypt_file}
\alias{decrypt_file}
\alias{encrypt_file}
\title{Encrypt and decrypt a file}
\usage{
encrypt_file(src, dest, config)

decrypt_file(src, dest, config)
}
\arguments{
\item{src}{The source filename (clear file for
\code{encrypt_file}, encrypted file for \code{decrypt_file}).}

\item{dest}{The destination filename (encrypted file for
\code{encrypt_file}, decrypted file for \code{decrypt_file}).
Can be the same as \code{src} in which case the file will be
overwritten without warning.}

\item{config}{A \code{encryptr_config} object describing the
encryption approach to use.}
}
\description{
Encrypt and decrypt a file.
}

