% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_symmetric}
\alias{config_authenticated}
\alias{config_public}
\alias{config_symmetric}
\title{Encryption configuration}
\usage{
config_symmetric(path_key)

config_public(path_key, path_pub)

config_authenticated(path_key, path_pub)
}
\arguments{
\item{path_key}{Path to, or contents of, the symmetric key
(\code{config_symmetric}) or your personal private key
(\code{config_public}, \code{config_authenticated}).}

\item{path_pub}{Path to, or contents of, the public key; your
personal public key (\code{config_public}) or the other party's
key (\code{config_authenticated}).}
}
\description{
Load keys for use with \code{encrypt}/\code{decrypt}
}
\details{
This interface will change because I'd like to push it into global
state with \emph{named} things that get pulled from a global cache
of options so that the wrappers can be as thin as possible.  But
for now, these are the full things and will probably always be
supported.
}

