% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openssl.R
\name{keypair_openssl}
\alias{keypair_openssl}
\title{Asymmetric encryption with openssl}
\usage{
keypair_openssl(pub, key, envelope = TRUE, password = NULL)
}
\arguments{
\item{pub}{An openssl public key.  Usually this will be the path
to the key, in which case it may either the path to a public key
or be the path to a directory containing a file
\code{id_rsa.pub}}

\item{key}{An openssl private key.  Usually this will be the path
to the key, in which case it may either the path to a private
key or be the path to a directory containing a file.  You may
specify \code{NULL} here, in which case the environment variable
\code{USER_KEY} is checked and if that is not defined then
\code{~/.ssh} will be used.}

\item{envelope}{A logical indicating if "envelope" encryption
functions should be used.  If so, then we use
\code{openssl::encrypt_envelope} and
\code{openssl::decrypt_envelope}.  If \code{FALSE} then we use
\code{openssl::rsa_encrypt} and \code{openssl::rsa_decrypt}.
See the openssl docs for further details.}

\item{password}{A password for the private key.  If \code{NULL}
then you will be prompted interactively for your password, and
if a string then that string will be used as the password (but
be careful in scripts!)}
}
\description{
Wrap a pair of openssl keys.  You should pass your private key and
the public key of the person that you are communicating with.
}
