% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sodium.R
\name{keypair_sodium}
\alias{keypair_sodium}
\title{Asymmetric encryption with sodium}
\usage{
keypair_sodium(pub, key, authenticated = TRUE)
}
\arguments{
\item{pub}{A sodium public key.  This is either a raw vector of
length 32 or a path to file containing the contents of the key
(written by \code{writeBin}.}

\item{key}{A sodium private key.  This is either a raw vector of
length 32 or a path to file containing the contents of the key
(written by \code{writeBin}.}

\item{authenticated}{Logical, indicating if authenticated
encryption (via \code{sodium::auth_encrypt} /
\code{sodium::auth_decrypt}) should be used.  If \code{FALSE}
then \code{sodium::simple_encrypt} /
\code{sodium::simple_decrypt} will be used.  The difference is
that with \code{authenticated = TRUE} the message is signed with
your private key so that tampering with the message will be
detectd.}
}
\description{
Wrap a pair of sodium keys for asymmetric encryption.  You should
pass your private key and the public key of the person that you
are communicating with.
}
\details{
\emph{NOTE}: the order here (pub, key) is very important; if the
wrong order is used you cannot decrypt things.  Unfortunately
because sodium keys are just bytesstrings there is nothing to
distinguish the public and private keys so this is a pretty easy
mistake to make.
}

