% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session_key_refresh}
\alias{session_key_refresh}
\title{Refresh the session key}
\usage{
session_key_refresh()
}
\description{
Refresh the session key, invalidating all keys created by
\code{\link{key_openssl}}, \code{\link{keypair_openssl}},
\code{\link{key_sodium}} and \code{\link{keypair_sodium}}.
}
\details{
Running this function will invalidate \emph{all} keys loaded with
the above functions.  It should not be needed very often.
}
