% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openssl.R
\name{config_openssl}
\alias{config_openssl}
\title{Load openssl keys}
\usage{
config_openssl(path = NULL, private = TRUE)
}
\arguments{
\item{path}{The path to the key.  If \code{NULL}, the environment
variables \code{USER_KEY} and \code{USER_PUBKEY} are checked,
and if they are empty then \code{~/.ssh/id_rsa} is used.  The
path can be either the full path to a public or private key, or
it can be the path to a directory containing the keypair (as
\code{id_rsa} and \code{id_rsa.pub}).}

\item{private}{A logical scalar indicating if the private key
should be loaded.  Because keypairs are often password
protected, and because public keys are useful on their own,
often it will not be needed to unlock and load the private key.}
}
\description{
Load keys for use with \code{\link{encrypt}} /
\code{\link{decrypt}}, using openssl encryption.  This is in
contrast to the sodium supported keys in (e.g.,
\code{\link{config_symmetric}} as here we assume that keys are
specified as the path to files as there are standard formats for
these.
}
\examples{
\dontrun{
cfg <- config_openssl(private=FALSE)
cfg$encrypt(charToRaw("hello"))
}
}

