% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{config_openssl}
\alias{config_openssl}
\title{Load openssl keys}
\usage{
config_openssl(public = NULL, private = TRUE, envelope = TRUE)
}
\arguments{
\item{public}{The path to the pulblic key.  If \code{NULL}, we first
check the R option \code{encryptr.user.path} and if that is
unset check the environment variable \code{USER_PUBKEY}, and if
that is unset default to \code{~/.ssh/id_rsa.pub}.  If any of
these are directories, we look for a file \code{id_rsa.pub}
within that directory.  If the file does not exist it will raise
an error.}

\item{private}{One of (1) A logical \code{FALSE} indicating that a
private ey should not be loaded, (2) A logical \code{TRUE}
indicating that the location of the private key should be
determined from the location of the public key, (3) \code{NULL}
to infer the location of the private key from the
\code{encryptr.user.path} R variable and \code{USER_KEY}
environment variable (in a similar way to \code{public} above).
Because keypairs are often password protected, and because
public keys are useful on their own, often it will not be needed
to unlock and load the private key.}

\item{envelope}{A logical scalar indicating if the "envelope"
functions should be used.  If \code{TRUE}, then this is used for
messaging (via \code{\link{encrypt_envelope}}).  This is what
you would use to communicate with a third party using shared
public keys and non- private keys.  If \code{FALSE}, then
\code{\link{rsa_encrypt}} is used.  This is less useful because
it can only encrypt messages smaller than the size of the key.}
}
\description{
Load an openssl key pair for use with \code{\link{encrypt}} /
\code{\link{decrypt}}.  This is for use with \emph{asymmetric
public key encryption} (see \code{vignette("encryptr")} and the
\code{openssl} package documentation for more details).  As such
there are a couple of different ways of loading the keys.
}
\details{
The default will load your own public and private key which is
useful for decrypting message sent to you.  If you want to send
someone else a message you will need to load \emph{their} public
key, in which case you would use

\preformatted{encryptr::config_openssl(path_pub, FALSE)}

The \code{FALSE} prevents loading a private key (as you do not
have the private key associated with that public key.

This function is designed in a different way to the similar sodium
functions (e.g., \code{\link{config_sodium_symmetric}}) as here we
assume that keys are specified as the path to files as there are
standard formats for these.  There are many different places that
keys may be (especially when loading someone else's public keys)
so the logic here is a little convoluted.
}
\examples{
\dontrun{
cfg <- config_openssl(private=FALSE)
secret <- encrypt_string("hello", NULL, cfg)
decrypt_string(secret, cfg)
}
}

