% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_request_access}
\alias{data_request_access}
\alias{data_key}
\title{User commands}
\usage{
data_request_access(path_data = ".", path_user = NULL, quiet = FALSE)

data_key(path_data = ".", path_user = NULL, test = TRUE,
  quiet = FALSE)
}
\arguments{
\item{path_data}{Path to the data.  The default is the current
working directory.}

\item{path_user}{Path to the directory with your user key.
Usually this can be omitted.  Use the \code{cyphr.user.path}
global option (i.e., via \code{options()}) to set this more
conveniently.}

\item{quiet}{Suppress printing of informative messages.}

\item{test}{Test that the encryption is working?  (Recommended)}
}
\description{
User commands
}
