% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{config_symmetric}
\alias{config_authenticated}
\alias{config_public}
\alias{config_symmetric}
\title{Encryption configuration}
\usage{
config_symmetric(path_key)

config_public(path_pub, path_key)

config_authenticated(path_pub, path_key)
}
\arguments{
\item{path_key}{Path to, or contents of, the symmetric key
(\code{config_symmetric}) or your personal private key
(\code{config_public}, \code{config_authenticated}).  For
\code{config_public}, this may be \code{NULL} in which case only
encryption can be carried out.}

\item{path_pub}{Path to, or contents of, the public key; your
personal public key (\code{config_public}) or the other party's
key (\code{config_authenticated}).}
}
\description{
Load sodium keys for use with \code{encrypt}/\code{decrypt}.
}
\details{
Note that the interface here is quite different to that for
\code{\link{config_openssl}} which loads ssh keys; here we expect
keys to be provided as raw data, or as links to binary files that
contain that raw data.  In contrast for openssl we use standard
key formats.
}

