% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openssl.R
\name{ssh_keygen}
\alias{ssh_keygen}
\title{Wrapper around ssh-keygen}
\usage{
ssh_keygen(path = tempfile(), password = TRUE)
}
\arguments{
\item{path}{A directory in which to create a keypair}

\item{password}{The password for the key.  The default will prompt
interactively (but without echoing the password).  Other valid
options are \code{FALSE} (no password), a string or a function
that will take a single argument "prompt" and return a string
password (as with the openssl package).}
}
\description{
Wrapper around ssh-keygen(1).
}

